requirements.txt
flask>=3.1.0
markupsafe
nltk==3.8.1
deep-translator>=1.11.0
geopy>=2.4.0
spotipy>=2.23.0
transformers>=4.36.0
torch>=2.2.0
numpy>=1.21.0,<2.0.0
wordnet
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
beautifulsoup4>=4.13.0
emoji>=2.8.0
huggingface-hub>=0.30.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
scikit-learn>=1.3.0
tqdm>=4.66.0
werkzeug>=3.0.0